feat(frontend): add forgot password modal
