+++ frontend/src/components/Header.vue
@@ -5,8 +5,9 @@ import { Dialog } from '@/components/ui/dialog'
 import { Button } from '@/components/ui/button'
 import LoginForm from '@/components/LoginForm.vue'
 import RegisterForm from '@/components/RegisterForm.vue'
+import ForgotPasswordForm from '@/components/ForgotPasswordForm.vue'
 
-const mode = ref<'login' | 'register'>('login')
+const mode = ref<'login' | 'register' | 'forgot'>('login')
 </script>
 
 <template>
@@ -22,11 +23,17 @@ const mode = ref<'login' | 'register'>('login')
         </Button>
       </template>
       <template #title>
-        {{ mode === 'login' ? 'Log in' : 'Sign up' }}
+        {{
+          mode === 'login'
+            ? 'Log in'
+            : mode === 'register'
+              ? 'Sign up'
+              : 'Forgot password'
+        }}
       </template>
       <component
-        :is="mode === 'login' ? LoginForm : RegisterForm"
-        @switch="(m: 'login' | 'register') => (mode = m)"
+        :is="mode === 'login' ? LoginForm : mode === 'register' ? RegisterForm : ForgotPasswordForm"
+        @switch="(m: 'login' | 'register' | 'forgot') => (mode = m)"
       />
     </Dialog>
   </header>
